feat(room): add Arcjet protection for joining rooms
This commit adds Arcjet protection for the "join-room" socket event on the server-side. The changes include:

- Integrating Arcjet with the server setup, including token bucket and shield rules.
- Checking the Arcjet decision before allowing the client to join a room.
- Returning a structured response object to the client with success/error information.
- Updating the client-side `RoomJoinForm` component to handle the new response format.
- Adding a new `TJoinRoomCallback` type to represent the server's response.

The main goal of these changes is to protect the server from potential abuse or misuse of the "join-room" functionality, ensuring a more secure and reliable real-time communication experience for the application.
