+++ server/server.ts
@@ -1,6 +1,25 @@
 import { Server } from "socket.io";
 import { createServer } from "node:http";
 import { onDisconnect, onHello } from "@/server/functions";
+import arcjet, { shield, tokenBucket } from "@arcjet/node";
+import { Logger } from "next-axiom";
+import { TJoinRoomCallback } from "@/types/room";
+
+const aj = arcjet({
+    key: process.env.ARCJET_KEY!,
+    characteristics: ["socket"],
+    rules: [
+        tokenBucket({
+            mode: "LIVE",
+            refillRate: 1,
+            interval: 6,
+            capacity: 5,
+        }),
+        shield({
+            mode: "LIVE",
+        }),
+    ],
+});
 
 const server = createServer();
 const io = new Server(server, {
@@ -14,7 +33,28 @@ io.on("connection", (socket) => {
     socket.on("disconnect", onDisconnect);
     socket.on("hello", onHello);
 
-    socket.on("join-room", async (roomName: string, callback: (response: string) => void) => {
+    socket.on("join-room", async (roomName: string, callback: (response: TJoinRoomCallback) => void) => {
+        const log = new Logger({ source: "Join Room" });
+        const decision = await aj.protect(socket, { requested: 1, socket: socket.id });
+
+        if (decision.isDenied()) {
+            log.info(JSON.stringify(decision.reason));
+            callback({ success: false, errorMessage: "You can't join" });
+            return;
+        }
+
+        if (decision.isErrored()) {
+            log.error(JSON.stringify(decision.reason));
+            callback({ success: false, errorMessage: "Something went wrong" });
+            return;
+        }
+
+        if (decision.isChallenged()) {
+            log.warn(JSON.stringify(decision.reason));
+            callback({ success: false, errorMessage: "You can't join" });
+            return;
+        }
+
         socket.rooms.forEach((room) => {
             socket.leave(room);
         });
@@ -30,7 +70,7 @@ io.on("connection", (socket) => {
 
         console.log(`User ${socket.id} joined room ${roomName}`);
 
-        callback(roomName);
+        callback({ success: true, roomName });
     });
 });
