+++ types/room.ts
@@ -1,0 +1,5 @@
+export type TJoinRoomCallback = {
+    success: boolean;
+    errorMessage?: string;
+    roomName?: string;
+};
